Module Web.Response

type body

The type for request bodies.contents A body is a function that yields byte chunks of the request body as Some (bytes, pos, len) values. The bytes value must not be modified and is readable from pos to pos+len until the next call to the function. The function returns None at the end of stream.

val body : resp -> body

body r is r's body.

val status : resp -> status

status r is r's headers.